+++ intelmq/tests/lib/test_message.py
@@ -607,6 +607,13 @@ class TestMessageFactory(unittest.TestCase):
         with self.assertRaises(exceptions.InvalidValue):
             event.update({'source.asn': 'AS1'})
 
+    def test_invalid_harm_key(self):
+        """ Test if error is raised when using an invalid key. """
+        with self.assertRaises(exceptions.InvalidKey):
+            message.Event(harmonization={'event': {'foo..bar': {}}})
+        with self.assertRaises(exceptions.InvalidKey):
+            message.Event(harmonization={'event': {'foo.bar.': {}}})
+
 
 if __name__ == '__main__':  # pragma: no cover  # pragma: no cover
     unittest.main()